backend/requirements.txt
django
django-rest-framework
gunicorn
psycopg2-binary